realtek-poe: Make debug frames more descriptive
When running `realtek-poe -d`, it is helpful to know what the frame
types being discussed between the firmware and `realtek-poe` actually
mean. This commit adds a description after each frame to tell us what
type of frame is being sent so we do not have to manually decode it.

As commented-out in the change, this code was generated using the
following shell pipeline:

```sh
curl https://svanheule.net/switches/software/broadcom_poe_control_protocol |
	 grep -m1 Set.port.enable -A 200 |
	 sed 's/<[^>]*>//g' |
	 tr -s ' ' |
	 grep -P '^\t' |
	 grep ' .. ' |
	 grep -vw 'ID' |
	 sed -E 's/^\s*//;s/([^ ]*) (.*)/\t\tif (data[0] == 0x\1)\n\t\t\tfprintf(stderr, "\2\\n");/'
```