realtek-poe: Fix out-of-bounds access in poe_reply_consume()
Imagine we get a packet from the PoE controller over the serial port.
Everything is running fine. All of a sudden realtek-poe crashes
because some numbnut (you) used a fixed-size array as a map, and never
checked the bounds.

This can be easily fixed by checking that:
    reply[0] > ARRAY_SIZE(reply_handler)
before dereferencing the reply_handler array.